6-size.c

+++ 0x00-hello_world/6-size.c
@@ -11,11 +11,11 @@ long int c;
 long long int d;
 float e;
 
-printf("Size of a char= %zu byte(s)", sizeof(a));
-printf("\nSize of a int= %zu byte(s)", sizeof(b));
-printf("\nSize of a long int= %zu byte(s)", sizeof(c));
-printf("\nSize of a long long int= %zu byte(s)", sizeof(d));
-printf("\nSize of a float= %zu byte(s)", sizeof(e));
+printf("Size of a char= %zu byte(s)", (unsigned long)sizeof(a));
+printf("\nSize of a int= %zu byte(s)", (unsigned long)sizeof(b));
+printf("\nSize of a long int= %zu byte(s)", (unsigned long)sizeof(c));
+printf("\nSize of a long long int= %zu byte(s)", (unsigned long)sizeof(d));
+printf("\nSize of a float= %zu byte(s)", (unsigned long)sizeof(e));
 
 return (0);
 }
